Add another server restart to coordination_SUITE:recover_from_checkpoint
This ensures that we can recover from a snapshot which was promoted from
a checkpoint. The restart earlier in the test case ensures that we can
recover from a checkpoint but that doesn't provide an insights about
checkpoint promotion.

Without the fix for `ra_lib:sync_file/1` in the parent commit, this case
fails on the new calls to `ra:restart_server/2`.