Try a different checkpoint thinning approach
Instead of setting a fixed upper limit the max_checkpoints ra_log
settings instead determins the point at which thinning will beging
to be applied. This means there is no upper limit on the number
of checkpoints that can exist but the differnce in indexes will
begin to increase as the number goes up.