Be more lenient during pre init.
Ra currently expects that any registered server in names.dets also
has a valid directory. As these two not updated in a transaction it is
possible for the two to diverge hence this change introduce more leniency
in that a registered server where there is no directory for is not considered
an error that will stop the ra system from starting.