ra_snapshot: Defer validation for older checkpoints
This refactors `ra_snapshot:find_checkpoints/1` to cut down on some
work when there are many checkpoints. We scan through the checkpoint
directories to find the first (latest) valid checkpoint we can use for
recovery. Then we can defer using the `ra_snapshot:validate/1` callback
(which can be somewhat expensive) for any older checkpoints.

We assume that any checkpoints older than the latest valid checkpoint
are valid. We expect that invalid checkpoints would be created when a
machine terminates hard and unexpectedly and may stop an in-progress
write or leave a checkpoint file unsynced. This should only affect some
number of the latest checkpoints though. Once we've found a checkpoint
file that is valid, checkpoints older than that should be fully written
and synchronized too.

We also bail out of the search when we find a checkpoint that has a
lower index than the current snapshot index. Those checkpoints cannot
be promoted and should be deleted. We scan through the checkpoints from
most recent to least recent, so when we find a checkpoint with an older
index than the snapshot, we delete that checkpoint and any older
checkpoints.