Amend ra_machine:snapshot_installed/4
Instead of passing the plain old version we build a full snapshot meta
so that we can also compare index, term and membership as well as
the machine version.

Changed the order of arguments.

Removed the optional snapshot_installed/2 callback but retained
the fallback call in ra_machine in case it is still implemented
somwhere.