Add `counter_label` field to Ra server config
This new field sets the `seshat:label()` when creating a new counter for
a server (via `ra_counters:new/3` and in turn `seshat:new/4`).

This also deprecates the `counter` field. Having the creator of a Ra
server pass a `counters:counter_ref()` is unergonomic because the caller
must know the `seshat:fields_spec()` with which the counter should be
created.

Also in this change we remove the call to `ra_counters:new/2` from
the `ra_server_proc:config_defaults()` helper. By calling that function
when creating a default, the server process registered a counter
(causing an insertion into an ETS table in seshat) for the server even
if the caller specified a `counter` in the config.