Refactor consistent query
Instead of each member maintaining a shared query_index updated
by the heartbeat_rpc messages this commit changes the query_index
to just maintain each member's local sequence that is only used when
the member is a leader.

This reduces some complexity in the algorithm as possibly avoids
certain bugs.