Extend WAL and segment file names from 8 to 16 characters
In order to avoid running out of characters after a few years of
consistent usage.

The renames happens when the WAL and segment writer initialises
respectively. The segment writer writes a marker after a successful
upgrade in order to avoid listing files in all server directories
every time. The WAL always checks the current set of wal files which
should be ok given there is rarely more than one.