Fix log divergence bug.
When a follower with uncommitted entries in a prior term when it
was leader receives certain append entry rpc combinations it would
incorrectly increment and evalaute the leader's commit index and
in some cases apply uncommitted entries.

This commit ensures that the commit index is only incremented and
evaluated _if_ the local log's last index is validated against
the leaders's log.

Also handle the case where a snapshot with an index lower than
the follower's last applied index would cause the replication
to this follower to stop.