Don't allow a non-voter to start elections
When transfer_leadership is requested, the leader checks whether
the desired new leader is a voter and refuses to proceeed if it is not.
However, the desired new leader should also refuse to proceed,
if it doesn't consider itself a voter (the view of the membership
may be different on different nodes or might have changed since
the old/current leader checked it).

This fixes a behaviour we observed in testing where a newly added
member, which was still catching up, was picked to be a new leader
and actually became a leader, even though it was still in the `promotable` state.