ra_server: Ignore `#info_reply{}` from a node that is not part of cluster
[Why]
Between the time a node asks for another node's information and the time
it gets the reply, it is possible that this other node leaves the
cluster for whatever reason.

This causes a crash of the Ra server that asked the information because
it accesses the `cluster` map to get the state of the node that emitted
the `#info_reply{}` without checking first if the node is still part of
the cluster (i.e., there is a key corresponding to that node in the
`cluster` map).

To be exact, line 3651 raises a `badkey` exception:

    PeerState0 = maps:get(PeerId, Cluster),

[How]
We simply ignore the `#info_reply{}` from a node that is not present in
the `cluster` map.