Fix type mixups of ra_index() and ra_term()
This is a cosmetic change since the dialyzer isn't smart enough yet to
notice the mixup. With the addition of nominal types in OTP 28 though we
will eventually be able to get the dialyzer to notice this.

Note that in practice `#request_vote_rpc{}` and `#pre_vote_rpc{}` are
being constructed correctly with `ra_term()`s in these fields, so this
isn't a bug fix.