New server metadata persistenace
Until now, we persisted metadata for all servers in a given
Ra system in a single DETS file, owned by ra_log_meta gen batch server.
This was a source of bottlenecks, so we move each servers metadata
to a simple binary file. If the file doesn't exist, we try to
fetch metadata from ra_log_meta and immediately persist it, so that
we don't need ra_log_meta going forward.