ra_log: Allow resetting last-written index to snapshot
A leader might've taken a snapshot at the last index before a partition
takes place. It can then write commands after the partition is in effect
and the commands will not be committed. Then another leader can be
elected and ask the deposed leader to reset its log down to the snapshot
index. Previously this crashed since `ra_log:set_last_index/2` did not
consider the snapshot state.