Fix bug where the persisted last applied could be higher than last written.
This could happen as some members can apply indexes even if they haven't
yet been locally written fully as long as the commit index is higher.

In this case it would persist the last applied index even if the
entry had not been persisted by the WAL. If the WAL then crashed and
the node then restarted the member would fail at recovery as it would
try to recover up to a last applied index that does not exist.